Yet more charset changes.
"Wide" strings based on wchar_t being called UCS-2 is incorrect, it is platform dependent as to if they are UTF-16 or UCS-4. So renamed the function calls and documentation.

The above showed some errros to dow ith charset in the multi-part MIME support, so that was fixed along the way.

The old AsUCS2() function has been kept but marked deprecated.